test: move controlplane endpoint unit tests
